+++ languages/de/tl_fh-counter.php
@@ -1,0 +1,35 @@
+<?php
+
+/**
+ * Contao Open Source CMS
+ *
+ * Copyright (c) 2005-2014 Leo Feyer
+ *
+ * @package   fh-counter
+ * @author    Frank Hoppe
+ * @license   GNU/LGPL
+ * @copyright Frank Hoppe 2014
+ */
+
+
+/**
+ * Fields
+ */
+$GLOBALS['TL_LANG']['tl_fh-counter']['title'] = array('Title', 'Please enter the title.');
+
+
+/**
+ * Legends
+ */
+$GLOBALS['TL_LANG']['tl_fh-counter']['title_legend'] = 'Title';
+
+
+/**
+ * Buttons
+ */
+$GLOBALS['TL_LANG']['tl_fh-counter']['new']    = array('New ', 'Create a new ');
+$GLOBALS['TL_LANG']['tl_fh-counter']['show']   = array(' details', 'Show the details of  ID %s');
+$GLOBALS['TL_LANG']['tl_fh-counter']['edit']   = array('Edit ', 'Edit  ID %s');
+$GLOBALS['TL_LANG']['tl_fh-counter']['cut']    = array('Move ', 'Move  ID %s');
+$GLOBALS['TL_LANG']['tl_fh-counter']['copy']   = array('Duplicate ', 'Duplicate  ID %s');
+$GLOBALS['TL_LANG']['tl_fh-counter']['delete'] = array('Delete ', 'Delete  ID %s');