+++ languages/en/default.php
@@ -1,0 +1,18 @@
+<?php
+
+/**
+ * Contao Open Source CMS
+ *
+ * Copyright (c) 2005-2014 Leo Feyer
+ *
+ * @package   fh-counter
+ * @author    Frank Hoppe
+ * @license   GNU/LGPL
+ * @copyright Frank Hoppe 2014
+ */
+
+
+/**
+ * Miscellaneous
+ */
+$GLOBALS['TL_LANG']['MSC'][''] = '';