+++ languages/en/modules.php
@@ -1,0 +1,24 @@
+<?php
+
+/**
+ * Contao Open Source CMS
+ *
+ * Copyright (c) 2005-2014 Leo Feyer
+ *
+ * @package   fh-counter
+ * @author    Frank Hoppe
+ * @license   GNU/LGPL
+ * @copyright Frank Hoppe 2014
+ */
+
+
+/**
+ * Back end modules
+ */
+$GLOBALS['TL_LANG']['MOD'][''] = array('', '');
+
+
+/**
+ * Front end modules
+ */
+$GLOBALS['TL_LANG']['FMD'][''] = array('', '');